Ausgaben mit printf verschönert, Methode aus Klasse Frontend importiert
